Remove globalThis.__lazilyLoadLinks feature flag
