+++ packages/host/app/routes/module.ts
@@ -93,11 +93,6 @@ export default class ModuleRoute extends Route<Model> {
   #releaseTimerBlock: (() => void) | undefined;
 
   deactivate() {
-    // Only unset lazilyLoadLinks if we last set it
-    if ((globalThis as any).__lazilyLoadLinksOwner === 'module') {
-      (globalThis as any).__lazilyLoadLinks = undefined;
-      (globalThis as any).__lazilyLoadLinksOwner = undefined;
-    }
     (globalThis as any).__boxelRenderContext = undefined;
     this.lastStoreResetKey = undefined;
     this.#authGuard.unregister();
@@ -111,8 +106,6 @@ export default class ModuleRoute extends Route<Model> {
     await super.beforeModel?.(transition);
     // activate() doesn't run early enough for this to be set before the model()
     // hook is run
-    (globalThis as any).__lazilyLoadLinks = true;
-    (globalThis as any).__lazilyLoadLinksOwner = 'module';
     (globalThis as any).__boxelRenderContext = true;
     this.#authGuard.register();
     if (!isTesting()) {
