+++ packages/host/app/routes/render.ts
@@ -104,11 +104,6 @@ export default class RenderRoute extends Route<Model> {
       currentURL: this.router.currentURL,
     });
     this.#setAllModelStatuses('unusable');
-    // Only unset lazilyLoadLinks if we last set it
-    if ((globalThis as any).__lazilyLoadLinksOwner === 'render') {
-      (globalThis as any).__lazilyLoadLinks = undefined;
-      (globalThis as any).__lazilyLoadLinksOwner = undefined;
-    }
     (globalThis as any).__boxelRenderContext = undefined;
   };
 
@@ -120,11 +115,6 @@ export default class RenderRoute extends Route<Model> {
   }
 
   deactivate() {
-    // Only unset lazilyLoadLinks if we last set it
-    if ((globalThis as any).__lazilyLoadLinksOwner === 'render') {
-      (globalThis as any).__lazilyLoadLinks = undefined;
-      (globalThis as any).__lazilyLoadLinksOwner = undefined;
-    }
     (globalThis as any).__boxelRenderContext = undefined;
     (globalThis as any).__renderInstance = undefined;
     window.removeEventListener('error', this.errorHandler);
@@ -149,8 +139,6 @@ export default class RenderRoute extends Route<Model> {
     await super.beforeModel?.(transition);
     // activate() doesn't run early enough for this to be set before the model()
     // hook is run
-    (globalThis as any).__lazilyLoadLinks = true;
-    (globalThis as any).__lazilyLoadLinksOwner = 'render';
     (globalThis as any).__boxelRenderContext = true;
     this.#authGuard.register();
     if (!isTesting()) {
