+++ packages/host/tests/acceptance/prerender-meta-test.gts
@@ -238,7 +238,6 @@ module('Acceptance | prerender | meta', function (hooks) {
   });
 
   hooks.afterEach(function () {
-    delete (globalThis as any).__lazilyLoadLinks;
     delete (globalThis as any).__boxelRenderContext;
   });
