+++ packages/host/tests/unit/garbage-collection-test.ts
@@ -41,7 +41,6 @@ module('Unit | identity-context garbage collection', function (hooks) {
   hooks.beforeEach(async function (this: RenderingTestContext) {
     loader = getService('loader-service').loader;
     api = await loader.import(`${baseRealm.url}card-api`);
-    delete (globalThis as any).__lazilyLoadLinks;
     delete (globalThis as any).__boxelRenderContext;
   });
