+++ packages/realm-server/tests/indexing-test.ts
@@ -1,47 +1,449 @@
-import { module, test, skip } from 'qunit';
+import { module, test } from 'qunit';
 import { dirSync } from 'tmp';
-import {
-  type DBAdapter,
-  type LooseSingleCardDocument,
-  type Realm,
-  type RealmPermissions,
-  type RealmAdapter,
-  CachingDefinitionLookup,
+import type {
+  DBAdapter,
+  LooseSingleCardDocument,
+  Realm,
+  RealmPermissions,
+  RealmAdapter,
 } from '@cardstack/runtime-common';
-import type { IndexedInstance } from '@cardstack/runtime-common';
+import type {
+  IndexedInstance,
+  QueuePublisher,
+  QueueRunner,
+} from '@cardstack/runtime-common';
 import {
-  createRealm,
-  testRealm,
   setupBaseRealmServer,
   setupDB,
   createVirtualNetwork,
   matrixURL,
-  setupPermissionedRealms,
   cleanWhiteSpace,
-  testRealmServerMatrixUserId,
   cardDefinition,
+  runTestRealmServer,
+  closeServer,
+  setupPermissionedRealms,
   cardInfo,
-  getTestPrerenderer,
 } from './helpers';
 import stripScopedCSSAttributes from '@cardstack/runtime-common/helpers/strip-scoped-css-attributes';
-import { basename } from 'path';
+import { join, basename } from 'path';
+import { resetCatalogRealms } from '../handlers/handle-fetch-catalog-realms';
+import type {
+  PgQueueRunner,
+  PgAdapter,
+  PgQueuePublisher,
+} from '@cardstack/postgres';
 
 function trimCardContainer(text: string) {
-  return cleanWhiteSpace(text).replace(
-    /<div .*? data-test-field-component-card>\s?[<!---->]*? (.*?) <\/div>/g,
-    '$1',
-  );
+  return cleanWhiteSpace(text)
+    .replace(/=""/g, '')
+    .replace(
+      /<div .*? data-test-field-component-card>\s?[<!---->]*? (.*?) <\/div>/g,
+      '$1',
+    );
 }
 
 let testDbAdapter: DBAdapter;
+const testRealm = new URL('http://127.0.0.1:4445/test/');
 
-// Using the node tests for indexing as it is much easier to support the dynamic
-// loading of cards necessary for indexing and the ability to manipulate the
-// underlying filesystem in a manner that doesn't leak into other tests (as well
-// as to test through loader caching)
+type TestRealmServerResult = Awaited<ReturnType<typeof runTestRealmServer>>;
+
+function makeTestRealmFileSystem(): Record<
+  string,
+  string | LooseSingleCardDocument
+> {
+  return {
+    'person.gts': `
+      import { contains, field, CardDef, Component } from "https://cardstack.com/base/card-api";
+      import StringField from "https://cardstack.com/base/string";
+      import NumberField from "https://cardstack.com/base/number";
+
+      export class Person extends CardDef {
+        @field firstName = contains(StringField);
+        @field hourlyRate = contains(NumberField);
+        static isolated = class Isolated extends Component<typeof this> {
+          <template>
+            <h1><@fields.firstName /> \${{@model.hourlyRate}}</h1>
+          </template>
+        }
+        static embedded = class Embedded extends Component<typeof this> {
+          <template>
+            <h1> Embedded Card Person: <@fields.firstName/></h1>
+
+            <style scoped>
+              h1 { color: red }
+            </style>
+          </template>
+        }
+        static fitted = class Fitted extends Component<typeof this> {
+          <template>
+            <h1> Fitted Card Person: <@fields.firstName/></h1>
+
+            <style scoped>
+              h1 { color: red }
+            </style>
+          </template>
+        }
+      }
+    `,
+    'pet-person.gts': `
+      import { contains, linksTo, field, CardDef, Component, StringField } from "https://cardstack.com/base/card-api";
+      import { Pet } from "./pet";
+
+      export class PetPerson extends CardDef {
+        @field firstName = contains(StringField);
+        @field pet = linksTo(() => Pet);
+        @field nickName = contains(StringField, {
+          computeVia: function (this: Person) {
+            if (this.pet?.firstName) {
+              return this.pet.firstName + "'s buddy";
+            }
+            return 'buddy';
+          },
+        });
+      }
+    `,
+    'pet.gts': `
+      import { contains, field, CardDef } from "https://cardstack.com/base/card-api";
+      import StringField from "https://cardstack.com/base/string";
+
+      export class Pet extends CardDef {
+        @field firstName = contains(StringField);
+      }
+    `,
+    'fancy-person.gts': `
+      import { contains, field, Component } from "https://cardstack.com/base/card-api";
+      import StringField from "https://cardstack.com/base/string";
+      import { Person } from "./person";
+
+      export class FancyPerson extends Person {
+        @field favoriteColor = contains(StringField);
+
+        static embedded = class Embedded extends Component<typeof this> {
+          <template>
+            <h1> Embedded Card Fancy Person: <@fields.firstName/></h1>
+
+            <style scoped>
+              h1 { color: pink }
+            </style>
+          </template>
+        }
+      }
+    `,
+    'post.gts': `
+      import { contains, field, linksTo, CardDef, Component } from "https://cardstack.com/base/card-api";
+      import StringField from "https://cardstack.com/base/string";
+      import { Person } from "./person";
+
+      export class Post extends CardDef {
+        static displayName = 'Post';
+        @field author = linksTo(Person);
+        @field message = contains(StringField);
+        static isolated = class Isolated extends Component<typeof this> {
+          <template>
+            <h1><@fields.message/></h1>
+            <h2><@fields.author/></h2>
+          </template>
+        }
+      }
+    `,
+    'boom.gts': `
+      import { contains, field, CardDef, Component } from "https://cardstack.com/base/card-api";
+      import StringField from "https://cardstack.com/base/string";
+
+      export class Boom extends CardDef {
+        @field firstName = contains(StringField);
+        static isolated = class Isolated extends Component<typeof this> {
+          <template>
+            <h1><@fields.firstName/>{{this.boom}}</h1>
+          </template>
+          get boom() {
+            throw new Error('intentional error');
+          }
+        }
+      }
+    `,
+    'boom2.gts': `
+      import { contains, field, CardDef, Component } from "https://cardstack.com/base/card-api";
+      import StringField from "https://cardstack.com/base/string";
+
+      export class Boom extends CardDef {
+        @field firstName = contains(StringField);
+        boom = () => {};
+        static isolated = class Isolated extends Component<typeof this> {
+          <template>
+            {{! From CS-7216 we are using a modifier in a strict mode template that is not imported }}
+            <h1 {{did-insert this.boom}}><@fields.firstName/></h1>
+          </template>
+        }
+      }
+    `,
+    'atom-boom.gts': `
+      import { contains, field, CardDef, Component } from "https://cardstack.com/base/card-api";
+      import StringField from "https://cardstack.com/base/string";
+
+      export class Boom extends CardDef {
+        @field firstName = contains(StringField);
+        static atom = class Atom extends Component<typeof this> {
+          <template>
+            <h1><@fields.firstName/>{{this.boom}}</h1>
+          </template>
+          get boom() {
+            throw new Error('intentional error');
+          }
+        }
+      }
+    `,
+    'embedded-boom.gts': `
+      import { contains, field, CardDef, Component } from "https://cardstack.com/base/card-api";
+      import StringField from "https://cardstack.com/base/string";
+
+      export class Boom extends CardDef {
+        @field firstName = contains(StringField);
+        static embedded = class Embedded extends Component<typeof this> {
+          <template>
+            <h1><@fields.firstName/>{{this.boom}}</h1>
+          </template>
+          get boom() {
+            throw new Error('intentional error');
+          }
+        }
+      }
+    `,
+    'fitted-boom.gts': `
+      import { contains, field, CardDef, Component } from "https://cardstack.com/base/card-api";
+      import StringField from "https://cardstack.com/base/string";
+
+      export class Boom extends CardDef {
+        @field firstName = contains(StringField);
+        static fitted = class Fitted extends Component<typeof this> {
+          <template>
+            <h1><@fields.firstName/>{{this.boom}}</h1>
+          </template>
+          get boom() {
+            throw new Error('intentional error');
+          }
+        }
+      }
+    `,
+    'mango.json': {
+      data: {
+        attributes: {
+          firstName: 'Mango',
+        },
+        meta: {
+          adoptsFrom: {
+            module: './person',
+            name: 'Person',
+          },
+        },
+      },
+    },
+    'vangogh.json': {
+      data: {
+        attributes: {
+          firstName: 'Van Gogh',
+          hourlyRate: 50,
+        },
+        meta: {
+          adoptsFrom: {
+            module: './person',
+            name: 'Person',
+          },
+        },
+      },
+    },
+    'hassan.json': {
+      data: {
+        attributes: {
+          firstName: 'Hassan',
+        },
+        relationships: {
+          pet: { links: { self: './ringo' } },
+        },
+        meta: {
+          adoptsFrom: {
+            module: './pet-person',
+            name: 'PetPerson',
+          },
+        },
+      },
+    },
+    'ringo.json': {
+      data: {
+        attributes: {
+          firstName: 'Ringo',
+        },
+        meta: {
+          adoptsFrom: {
+            module: './pet',
+            name: 'Pet',
+          },
+        },
+      },
+    },
+    'post-1.json': {
+      data: {
+        attributes: {
+          message: 'Who wants to fetch?!',
+        },
+        relationships: {
+          author: {
+            links: {
+              self: './vangogh',
+            },
+          },
+        },
+        meta: {
+          adoptsFrom: {
+            module: './post',
+            name: 'Post',
+          },
+        },
+      },
+    },
+    'bad-link.json': {
+      data: {
+        attributes: {
+          message: 'I have a bad link',
+        },
+        relationships: {
+          author: {
+            links: {
+              self: 'http://localhost:9000/this-is-a-link-to-nowhere',
+            },
+          },
+        },
+        meta: {
+          adoptsFrom: {
+            module: './post',
+            name: 'Post',
+          },
+        },
+      },
+    },
+    'boom.json': {
+      data: {
+        attributes: {
+          firstName: 'Boom!',
+        },
+        meta: {
+          adoptsFrom: {
+            module: './boom',
+            name: 'Boom',
+          },
+        },
+      },
+    },
+    'boom2.json': {
+      data: {
+        attributes: {
+          firstName: 'Boom!',
+        },
+        meta: {
+          adoptsFrom: {
+            module: './boom2',
+            name: 'Boom',
+          },
+        },
+      },
+    },
+    'atom-boom.json': {
+      data: {
+        attributes: {
+          firstName: 'Boom!',
+        },
+        meta: {
+          adoptsFrom: {
+            module: './atom-boom',
+            name: 'Boom',
+          },
+        },
+      },
+    },
+    'embedded-boom.json': {
+      data: {
+        attributes: {
+          firstName: 'Boom!',
+        },
+        meta: {
+          adoptsFrom: {
+            module: './embedded-boom',
+            name: 'Boom',
+          },
+        },
+      },
+    },
+    'fitted-boom.json': {
+      data: {
+        attributes: {
+          firstName: 'Boom!',
+        },
+        meta: {
+          adoptsFrom: {
+            module: './fitted-boom',
+            name: 'Boom',
+          },
+        },
+      },
+    },
+    'empty.json': {
+      data: {
+        attributes: {},
+        meta: {
+          adoptsFrom: {
+            module: 'https://cardstack.com/base/card-api',
+            name: 'CardDef',
+          },
+        },
+      },
+    },
+    'random-file.txt': 'hello',
+    'random-image.png': 'i am an image',
+    '.DS_Store':
+      'In  macOS, .DS_Store is a file that stores custom attributes of its containing folder',
+  };
+}
+
+async function startTestRealm({
+  dbAdapter,
+  publisher,
+  runner,
+}: {
+  dbAdapter: DBAdapter;
+  publisher: QueuePublisher;
+  runner: QueueRunner;
+}): Promise<TestRealmServerResult> {
+  let virtualNetwork = createVirtualNetwork();
+  let dir = dirSync().name;
+  let testRealmServer = await runTestRealmServer({
+    testRealmDir: dir,
+    realmsRootPath: join(dir, 'realm_server_1'),
+    virtualNetwork,
+    realmURL: testRealm,
+    dbAdapter: dbAdapter as PgAdapter,
+    publisher: publisher as PgQueuePublisher,
+    runner: runner as PgQueueRunner,
+    matrixURL,
+    fileSystem: makeTestRealmFileSystem(),
+  });
+  await testRealmServer.testRealm.start();
+  return testRealmServer;
+}
+
+async function stopTestRealm(testRealmServer?: TestRealmServerResult) {
+  if (!testRealmServer) {
+    return;
+  }
+  testRealmServer.testRealm.unsubscribe();
+  await closeServer(testRealmServer.testRealmHttpServer);
+  resetCatalogRealms();
+}
 
 module(basename(__filename), function () {
-  module('indexing', function (hooks) {
+  module('indexing (read only)', function (hooks) {
+    let realm: Realm;
+    let testRealmServer: TestRealmServerResult | undefined;
+
     async function getInstance(
       realm: Realm,
       url: URL,
@@ -53,392 +455,21 @@ module(basename(__filename), function () {
       return maybeInstance;
     }
 
-    let dir: string;
-    let realm: Realm;
-    let adapter: RealmAdapter;
-
-    setupBaseRealmServer(hooks, matrixURL, { disablePrerenderer: true });
+    setupBaseRealmServer(hooks, matrixURL);
 
     setupDB(hooks, {
-      beforeEach: async (dbAdapter, publisher, runner) => {
+      before: async (dbAdapter, publisher, runner) => {
         testDbAdapter = dbAdapter;
-        let virtualNetwork = createVirtualNetwork();
-        let prerenderer = await getTestPrerenderer();
-        let definitionLookup = new CachingDefinitionLookup(
-          dbAdapter,
-          prerenderer,
-          virtualNetwork,
-        );
-        dir = dirSync().name;
-        ({ realm, adapter } = await createRealm({
-          disablePrerenderer: true,
-          withWorker: true,
-          definitionLookup,
-          dir,
-          virtualNetwork,
+        testRealmServer = await startTestRealm({
           dbAdapter,
           publisher,
           runner,
-          fileSystem: {
-            'person.gts': `
-            import { contains, field, CardDef, Component } from "https://cardstack.com/base/card-api";
-            import StringField from "https://cardstack.com/base/string";
-            import NumberField from "https://cardstack.com/base/number";
-
-            export class Person extends CardDef {
-              @field firstName = contains(StringField);
-              @field hourlyRate = contains(NumberField);
-              static isolated = class Isolated extends Component<typeof this> {
-                <template>
-                  <h1><@fields.firstName /> \${{@model.hourlyRate}}</h1>
-                </template>
-              }
-
-              static head = class Head extends Component<typeof this> {
-                <template>
-                  <title>{{@model.firstName}}!</title>
-                </template>
-              }
-
-              static embedded = class Embedded extends Component<typeof this> {
-                <template>
-                  <h1> Embedded Card Person: <@fields.firstName/></h1>
-
-                  <style scoped>
-                    h1 { color: red }
-                  </style>
-                </template>
-              }
-              static fitted = class Fitted extends Component<typeof this> {
-                <template>
-                  <h1> Fitted Card Person: <@fields.firstName/></h1>
-
-                  <style scoped>
-                    h1 { color: red }
-                  </style>
-                </template>
-              }
-            }
-          `,
-            'pet-person.gts': `
-            import { contains, linksTo, field, CardDef, Component, StringField } from "https://cardstack.com/base/card-api";
-            import { Pet } from "./pet";
-
-            export class PetPerson extends CardDef {
-              @field firstName = contains(StringField);
-              @field pet = linksTo(() => Pet);
-              @field nickName = contains(StringField, {
-                computeVia: function (this: Person) {
-                  if (this.pet?.firstName) {
-                    return this.pet.firstName + "'s buddy";
-                  }
-                  return firstName;
-                },
-              });
-            }
-          `,
-            'pet.gts': `
-            import { contains, field, CardDef } from "https://cardstack.com/base/card-api";
-            import StringField from "https://cardstack.com/base/string";
-
-            export class Pet extends CardDef {
-              @field firstName = contains(StringField);
-            }
-          `,
-            'fancy-person.gts': `
-            import { contains, field, Component } from "https://cardstack.com/base/card-api";
-            import StringField from "https://cardstack.com/base/string";
-            import { Person } from "./person";
-
-            export class FancyPerson extends Person {
-              @field favoriteColor = contains(StringField);
-
-              static embedded = class Embedded extends Component<typeof this> {
-                <template>
-                  <h1> Embedded Card Fancy Person: <@fields.firstName/></h1>
-
-                  <style scoped>
-                    h1 { color: pink }
-                  </style>
-                </template>
-              }
-            }
-          `,
-            'post.gts': `
-            import { contains, field, linksTo, CardDef, Component } from "https://cardstack.com/base/card-api";
-            import StringField from "https://cardstack.com/base/string";
-            import { Person } from "./person";
-
-            export class Post extends CardDef {
-              static displayName = 'Post';
-              @field author = linksTo(Person);
-              @field message = contains(StringField);
-              static isolated = class Isolated extends Component<typeof this> {
-                <template>
-                  <h1><@fields.message/></h1>
-                  <h2><@fields.author/></h2>
-                </template>
-              }
-            }
-          `,
-            'boom.gts': `
-            import { contains, field, CardDef, Component } from "https://cardstack.com/base/card-api";
-            import StringField from "https://cardstack.com/base/string";
-
-            export class Boom extends CardDef {
-              @field firstName = contains(StringField);
-              static isolated = class Isolated extends Component<typeof this> {
-                <template>
-                  <h1><@fields.firstName/>{{this.boom}}</h1>
-                </template>
-                get boom() {
-                  throw new Error('intentional error');
-                }
-              }
-            }
-          `,
-            'boom2.gts': `
-            import { contains, field, CardDef, Component } from "https://cardstack.com/base/card-api";
-            import StringField from "https://cardstack.com/base/string";
-
-            export class Boom extends CardDef {
-              @field firstName = contains(StringField);
-              boom = () => {};
-              static isolated = class Isolated extends Component<typeof this> {
-                <template>
-                  {{! From CS-7216 we are using a modifier in a strict mode template that is not imported }}
-                  <h1 {{did-insert this.boom}}><@fields.firstName/></h1>
-                </template>
-              }
-            }
-          `,
-            'atom-boom.gts': `
-            import { contains, field, CardDef, Component } from "https://cardstack.com/base/card-api";
-            import StringField from "https://cardstack.com/base/string";
-
-            export class Boom extends CardDef {
-              @field firstName = contains(StringField);
-              static atom = class Atom extends Component<typeof this> {
-                <template>
-                  <h1><@fields.firstName/>{{this.boom}}</h1>
-                </template>
-                get boom() {
-                  throw new Error('intentional error');
-                }
-              }
-            }
-          `,
-            'embedded-boom.gts': `
-            import { contains, field, CardDef, Component } from "https://cardstack.com/base/card-api";
-            import StringField from "https://cardstack.com/base/string";
-
-            export class Boom extends CardDef {
-              @field firstName = contains(StringField);
-              static embedded = class Embedded extends Component<typeof this> {
-                <template>
-                  <h1><@fields.firstName/>{{this.boom}}</h1>
-                </template>
-                get boom() {
-                  throw new Error('intentional error');
-                }
-              }
-            }
-          `,
-            'fitted-boom.gts': `
-            import { contains, field, CardDef, Component } from "https://cardstack.com/base/card-api";
-            import StringField from "https://cardstack.com/base/string";
-
-            export class Boom extends CardDef {
-              @field firstName = contains(StringField);
-              static fitted = class Fitted extends Component<typeof this> {
-                <template>
-                  <h1><@fields.firstName/>{{this.boom}}</h1>
-                </template>
-                get boom() {
-                  throw new Error('intentional error');
-                }
-              }
-            }
-          `,
-            'mango.json': {
-              data: {
-                attributes: {
-                  firstName: 'Mango',
-                },
-                meta: {
-                  adoptsFrom: {
-                    module: './person',
-                    name: 'Person',
-                  },
-                },
-              },
-            },
-            'vangogh.json': {
-              data: {
-                attributes: {
-                  firstName: 'Van Gogh',
-                  hourlyRate: 50,
-                },
-                meta: {
-                  adoptsFrom: {
-                    module: './person',
-                    name: 'Person',
-                  },
-                },
-              },
-            },
-            'hassan.json': {
-              data: {
-                attributes: {
-                  firstName: 'Hassan',
-                },
-                relationships: {
-                  pet: { links: { self: './ringo' } },
-                },
-                meta: {
-                  adoptsFrom: {
-                    module: './pet-person',
-                    name: 'PetPerson',
-                  },
-                },
-              },
-            },
-            'ringo.json': {
-              data: {
-                attributes: {
-                  firstName: 'Ringo',
-                },
-                meta: {
-                  adoptsFrom: {
-                    module: './pet',
-                    name: 'Pet',
-                  },
-                },
-              },
-            },
-            'post-1.json': {
-              data: {
-                attributes: {
-                  message: 'Who wants to fetch?!',
-                },
-                relationships: {
-                  author: {
-                    links: {
-                      self: './vangogh',
-                    },
-                  },
-                },
-                meta: {
-                  adoptsFrom: {
-                    module: './post',
-                    name: 'Post',
-                  },
-                },
-              },
-            },
-            'bad-link.json': {
-              data: {
-                attributes: {
-                  message: 'I have a bad link',
-                },
-                relationships: {
-                  author: {
-                    links: {
-                      self: 'http://localhost:9000/this-is-a-link-to-nowhere',
-                    },
-                  },
-                },
-                meta: {
-                  adoptsFrom: {
-                    module: './post',
-                    name: 'Post',
-                  },
-                },
-              },
-            },
-            'boom.json': {
-              data: {
-                attributes: {
-                  firstName: 'Boom!',
-                },
-                meta: {
-                  adoptsFrom: {
-                    module: './boom',
-                    name: 'Boom',
-                  },
-                },
-              },
-            },
-            'boom2.json': {
-              data: {
-                attributes: {
-                  firstName: 'Boom!',
-                },
-                meta: {
-                  adoptsFrom: {
-                    module: './boom2',
-                    name: 'Boom',
-                  },
-                },
-              },
-            },
-            'atom-boom.json': {
-              data: {
-                attributes: {
-                  firstName: 'Boom!',
-                },
-                meta: {
-                  adoptsFrom: {
-                    module: './atom-boom',
-                    name: 'Boom',
-                  },
-                },
-              },
-            },
-            'embedded-boom.json': {
-              data: {
-                attributes: {
-                  firstName: 'Boom!',
-                },
-                meta: {
-                  adoptsFrom: {
-                    module: './embedded-boom',
-                    name: 'Boom',
-                  },
-                },
-              },
-            },
-            'fitted-boom.json': {
-              data: {
-                attributes: {
-                  firstName: 'Boom!',
-                },
-                meta: {
-                  adoptsFrom: {
-                    module: './fitted-boom',
-                    name: 'Boom',
-                  },
-                },
-              },
-            },
-            'empty.json': {
-              data: {
-                attributes: {},
-                meta: {
-                  adoptsFrom: {
-                    module: 'https://cardstack.com/base/card-api',
-                    name: 'CardDef',
-                  },
-                },
-              },
-            },
-            'random-file.txt': 'hello',
-            'random-image.png': 'i am an image',
-            '.DS_Store':
-              'In  macOS, .DS_Store is a file that stores custom attributes of its containing folder',
-          },
-        }));
-        await realm.start();
+        });
+        realm = testRealmServer.testRealm;
+      },
+      after: async () => {
+        await stopTestRealm(testRealmServer);
+        testRealmServer = undefined;
       },
     });
 
@@ -478,13 +509,6 @@ module(basename(__filename), function () {
           cleanWhiteSpace(`<h1> Mango $</h1>`),
           'pre-rendered isolated format html is correct',
         );
-
-        assert.strictEqual(
-          trimCardContainer(stripScopedCSSAttributes(entry!.headHtml!)),
-          cleanWhiteSpace(`<title>Mango!</title>`),
-          'pre-rendered head format html is correct',
-        );
-
         assert.strictEqual(
           trimCardContainer(
             stripScopedCSSAttributes(
@@ -494,6 +518,21 @@ module(basename(__filename), function () {
           cleanWhiteSpace(`<h1> Embedded Card Person: Mango </h1>`),
           'pre-rendered embedded format html is correct',
         );
+
+        assert.ok(entry.headHtml, 'pre-rendered head format html is present');
+
+        let cleanedHead = cleanWhiteSpace(entry.headHtml!);
+
+        // TODO: restore in CS-9807
+        // assert.ok(
+        //   cleanedHead.includes('<title data-test-card-head-title>'),
+        //   `head html includes title: ${cleanedHead}`,
+        // );
+        assert.ok(
+          cleanedHead.includes(`property="og:url" content="${testRealm}mango"`),
+          `head html includes canonical url: ${cleanedHead}`,
+        );
+
         assert.strictEqual(
           trimCardContainer(
             stripScopedCSSAttributes(
@@ -515,7 +554,7 @@ module(basename(__filename), function () {
       if (entry?.type === 'error') {
         assert.strictEqual(
           entry.error.errorDetail.message,
-          'Encountered error rendering HTML for card: intentional error',
+          'intentional error',
         );
         assert.deepEqual(entry.error.errorDetail.deps, [
           `${testRealm}atom-boom`,
@@ -532,7 +571,7 @@ module(basename(__filename), function () {
       if (entry?.type === 'error') {
         assert.strictEqual(
           entry.error.errorDetail.message,
-          'Encountered error rendering HTML for card: intentional error',
+          'intentional error',
         );
         assert.deepEqual(entry.error.errorDetail.deps, [
           `${testRealm}embedded-boom`,
@@ -549,7 +588,7 @@ module(basename(__filename), function () {
       if (entry?.type === 'error') {
         assert.strictEqual(
           entry.error.errorDetail.message,
-          'Encountered error rendering HTML for card: intentional error',
+          'intentional error',
         );
         assert.deepEqual(entry.error.errorDetail.deps, [
           `${testRealm}fitted-boom`,
@@ -559,7 +598,7 @@ module(basename(__filename), function () {
       }
     });
 
-    test('can recover from rendering a card that has a template error', async function (assert) {
+    test('rendering a card that has a template error does not affect indexing other instances', async function (assert) {
       {
         let entry = await realm.realmIndexQueryEngine.cardDocument(
           new URL(`${testRealm}boom`),
@@ -567,7 +606,7 @@ module(basename(__filename), function () {
         if (entry?.type === 'error') {
           assert.strictEqual(
             entry.error.errorDetail.message,
-            'Encountered error rendering HTML for card: intentional error',
+            'intentional error',
           );
           assert.deepEqual(entry.error.errorDetail.deps, [`${testRealm}boom`]);
         } else {
@@ -581,7 +620,7 @@ module(basename(__filename), function () {
         if (entry?.type === 'error') {
           assert.strictEqual(
             entry.error.errorDetail.message,
-            'Encountered error rendering HTML for card: Attempted to resolve a modifier in a strict mode template, but it was not in scope: did-insert',
+            'Attempted to resolve a modifier in a strict mode template, but it was not in scope: did-insert',
           );
           assert.deepEqual(entry.error.errorDetail.deps, [`${testRealm}boom2`]);
         } else {
@@ -639,13 +678,187 @@ module(basename(__filename), function () {
           entry.error.errorDetail.message,
           'unable to fetch http://localhost:9000/this-is-a-link-to-nowhere: fetch failed',
         );
-        assert.deepEqual(entry.error.errorDetail.deps, [
+        let actualDeps = (entry.error.errorDetail.deps ?? []).map((d) =>
+          d.endsWith('.json') ? d.slice(0, -5) : d,
+        );
+        let expectedDeps = [
           `${testRealm}post`,
           `http://localhost:9000/this-is-a-link-to-nowhere`,
-        ]);
+        ];
+        assert.deepEqual(actualDeps.sort(), expectedDeps.sort());
       } else {
         assert.ok('false', 'expected search entry to be an error document');
       }
+    });
+
+    // Note this particular test should only be a server test as the nature of
+    // the TestAdapter in the host tests will trigger the linked card to be
+    // already loaded when in fact in the real world it is not.
+    test('it can index a card with a contains computed that consumes a linksTo field', async function (assert) {
+      const hassanId = `${testRealm}hassan`;
+      let queryEngine = realm.realmIndexQueryEngine;
+      let hassan = await queryEngine.cardDocument(new URL(hassanId));
+      if (hassan?.type === 'doc') {
+        assert.deepEqual(
+          hassan.doc.data.attributes,
+          {
+            title: 'Untitled Card',
+            nickName: "Ringo's buddy",
+            firstName: 'Hassan',
+            description: null,
+            thumbnailURL: null,
+            cardInfo,
+          },
+          'doc attributes are correct',
+        );
+        assert.deepEqual(
+          hassan.doc.data.relationships,
+          {
+            pet: {
+              links: {
+                self: './ringo',
+              },
+            },
+            'cardInfo.theme': {
+              links: {
+                self: null,
+              },
+            },
+          },
+          'doc relationships are correct',
+        );
+      } else {
+        assert.ok(
+          false,
+          `search entry was an error: ${hassan?.error.errorDetail.message}`,
+        );
+      }
+
+      let hassanEntry = await getInstance(realm, new URL(`${testRealm}hassan`));
+      if (hassanEntry) {
+        assert.deepEqual(
+          hassanEntry.searchDoc,
+          {
+            id: hassanId,
+            pet: {
+              id: `${testRealm}ringo`,
+              title: 'Untitled Card',
+              firstName: 'Ringo',
+              cardInfo: {
+                theme: null,
+              },
+            },
+            nickName: "Ringo's buddy",
+            _cardType: 'PetPerson',
+            firstName: 'Hassan',
+            title: 'Untitled Card',
+            cardInfo: {
+              theme: null,
+            },
+          },
+          'searchData is correct',
+        );
+      } else {
+        assert.ok(false, `could not find ${hassanId} in the index`);
+      }
+    });
+
+    test('sets resource_created_at for modules and instances', async function (assert) {
+      let entry = (await realm.realmIndexQueryEngine.module(
+        new URL(`${testRealm}fancy-person.gts`),
+      )) as { resourceCreatedAt: number };
+
+      assert.ok(entry!.resourceCreatedAt, 'resourceCreatedAt is set');
+
+      entry = (await realm.realmIndexQueryEngine.instance(
+        new URL(`${testRealm}mango`),
+      )) as { resourceCreatedAt: number };
+
+      assert.ok(entry!.resourceCreatedAt, 'resourceCreatedAt is set');
+    });
+
+    test('sets urls containing encoded CSS for deps for a module', async function (assert) {
+      let entry = (await realm.realmIndexQueryEngine.module(
+        new URL(`${testRealm}fancy-person.gts`),
+      )) as { deps: string[] };
+
+      let assertCssDependency = (
+        deps: string[],
+        pattern: RegExp,
+        fileName: string,
+      ) => {
+        assert.true(
+          !!deps.find((dep) => pattern.test(dep)),
+          `css for ${fileName} is in the deps`,
+        );
+      };
+
+      let dependencies = [
+        {
+          pattern: /fancy-person\.gts.*\.glimmer-scoped\.css$/,
+          fileName: 'fancy-person.gts',
+        },
+        {
+          pattern: /\/person\.gts.*\.glimmer-scoped\.css$/,
+          fileName: 'person.gts',
+        },
+        {
+          pattern:
+            /cardstack.com\/base\/default-templates\/embedded\.gts.*\.glimmer-scoped\.css$/,
+          fileName: 'default-templates/embedded.gts',
+        },
+        {
+          pattern:
+            /cardstack.com\/base\/default-templates\/isolated-and-edit\.gts.*\.glimmer-scoped\.css$/,
+          fileName: 'default-templates/isolated-and-edit.gts',
+        },
+        {
+          pattern:
+            /cardstack.com\/base\/default-templates\/missing-template\.gts.*\.glimmer-scoped\.css$/,
+          fileName: 'default-templates/missing-template.gts',
+        },
+        {
+          pattern:
+            /cardstack.com\/base\/default-templates\/field-edit\.gts.*\.glimmer-scoped\.css$/,
+          fileName: 'default-templates/field-edit.gts',
+        },
+        {
+          pattern:
+            /cardstack.com\/base\/links-to-many-component.gts.*\.glimmer-scoped\.css$/,
+          fileName: 'links-to-many-component.gts',
+        },
+        {
+          pattern:
+            /cardstack.com\/base\/links-to-editor.gts.*\.glimmer-scoped\.css$/,
+          fileName: 'links-to-editor.gts',
+        },
+        {
+          pattern:
+            /cardstack.com\/base\/contains-many-component.gts.*\.glimmer-scoped\.css$/,
+          fileName: 'contains-many-component.gts',
+        },
+        {
+          pattern:
+            /cardstack.com\/base\/field-component.gts.*\.glimmer-scoped\.css$/,
+          fileName: 'field-component.gts',
+        },
+      ];
+
+      dependencies.forEach(({ pattern, fileName }) => {
+        assertCssDependency(entry.deps, pattern, fileName);
+      });
+    });
+
+    test('will not invalidate non-json/non-executable files', async function (assert) {
+      let deletedEntries = (await testDbAdapter.execute(
+        `SELECT url FROM boxel_index WHERE is_deleted = TRUE`,
+      )) as { url: string }[];
+
+      let deletedEntryUrls = deletedEntries.map((row) => row.url);
+
+      ['random-file.txt', 'random-image.png', '.DS_Store'].forEach((file) => {
+        assert.notOk(deletedEntryUrls.includes(file));
+      });
     });
 
     test('can make a definition entry in the index', async function (assert) {
@@ -1272,8 +1485,33 @@ module(basename(__filename), function () {
         assert.ok('false', 'expected entry to be a card def');
       }
     });
-    // we're going to delete this test soon and it doesn't work with prerender-based definition lookup
-    skip('can incrementally index updated instance', async function (assert) {
+  });
+
+  module('indexing', function (hooks) {
+    let realm: Realm;
+    let adapter: RealmAdapter;
+    let testRealmServer: TestRealmServerResult | undefined;
+
+    setupBaseRealmServer(hooks, matrixURL);
+
+    setupDB(hooks, {
+      beforeEach: async (dbAdapter, publisher, runner) => {
+        testDbAdapter = dbAdapter;
+        testRealmServer = await startTestRealm({
+          dbAdapter,
+          publisher,
+          runner,
+        });
+        realm = testRealmServer.testRealm;
+        adapter = testRealmServer.testRealmAdapter;
+      },
+      afterEach: async () => {
+        await stopTestRealm(testRealmServer);
+        testRealmServer = undefined;
+      },
+    });
+
+    test('can incrementally index updated instance', async function (assert) {
       await realm.write(
         'mango.json',
         JSON.stringify({
@@ -1298,19 +1536,10 @@ module(basename(__filename), function () {
         },
       });
       assert.strictEqual(result.length, 1, 'found updated document');
-      assert.deepEqual(
-        // we splat because despite having the same shape, the constructors are different
-        { ...realm.realmIndexUpdater.stats },
-        {
-          instancesIndexed: 1,
-          instanceErrors: 0,
-          moduleErrors: 0,
-          modulesIndexed: 0,
-          definitionErrors: 0,
-          definitionsIndexed: 0,
-          totalIndexEntries: 26,
-        },
-        'indexed correct number of files',
+      assert.strictEqual(
+        realm.realmIndexUpdater.stats.instancesIndexed,
+        1,
+        'indexed updated instance',
       );
     });
 
@@ -1328,20 +1557,6 @@ module(basename(__filename), function () {
           throw new Error('boom!');
         `,
       );
-      assert.deepEqual(
-        // we splat because despite having the same shape, the constructors are different
-        { ...realm.realmIndexUpdater.stats },
-        {
-          instancesIndexed: 0,
-          instanceErrors: 2,
-          moduleErrors: 2,
-          modulesIndexed: 0,
-          definitionErrors: 0,
-          definitionsIndexed: 0,
-          totalIndexEntries: 20,
-        },
-        'indexed correct number of files',
-      );
       let petDefinitionEntry =
         await realm.realmIndexQueryEngine.getOwnDefinition({
           module: `${testRealm}pet`,
@@ -1358,20 +1573,6 @@ module(basename(__filename), function () {
           // syntax error
           export class Intentionally Thrown Error {}
         `,
-      );
-      assert.deepEqual(
-        // we splat because despite having the same shape, the constructors are different
-        { ...realm.realmIndexUpdater.stats },
-        {
-          instancesIndexed: 0,
-          instanceErrors: 4, // 1 post, 2 persons, 1 bad-link post
-          moduleErrors: 3, // post, fancy person, person
-          modulesIndexed: 0,
-          definitionErrors: 0,
-          definitionsIndexed: 0,
-          totalIndexEntries: 11,
-        },
-        'indexed correct number of files',
       );
       let { data: result } = await realm.realmIndexQueryEngine.search({
         filter: {
@@ -1413,20 +1614,6 @@ module(basename(__filename), function () {
             @field firstName = contains(StringField);
           }
         `,
-      );
-      assert.deepEqual(
-        // we splat because despite having the same shape, the constructors are different
-        { ...realm.realmIndexUpdater.stats },
-        {
-          instancesIndexed: 3, // 1 post and 2 persons
-          instanceErrors: 1,
-          moduleErrors: 0,
-          modulesIndexed: 3,
-          definitionErrors: 0,
-          definitionsIndexed: 3, // Person card, Post card, FancyPerson card
-          totalIndexEntries: 20,
-        },
-        'indexed correct number of files',
       );
       result = (
         await realm.realmIndexQueryEngine.search({
@@ -1475,19 +1662,6 @@ module(basename(__filename), function () {
           }
         `,
       );
-      assert.deepEqual(
-        { ...realm.realmIndexUpdater.stats },
-        {
-          instancesIndexed: 0,
-          instanceErrors: 2,
-          moduleErrors: 2,
-          modulesIndexed: 0,
-          definitionErrors: 0,
-          definitionsIndexed: 0,
-          totalIndexEntries: 20,
-        },
-        'indexed correct number of files',
-      );
       let petDefinitionEntry =
         await realm.realmIndexQueryEngine.getOwnDefinition({
           module: `${testRealm}pet`,
@@ -1531,20 +1705,6 @@ module(basename(__filename), function () {
       );
 
       // Since the name is ready, the pet should be indexed and not in an error state
-      assert.deepEqual(
-        { ...realm.realmIndexUpdater.stats },
-        {
-          instancesIndexed: 1,
-          instanceErrors: 1,
-          moduleErrors: 0,
-          modulesIndexed: 3,
-          definitionErrors: 0,
-          definitionsIndexed: 3,
-          totalIndexEntries: 27,
-        },
-        'indexed correct number of files',
-      );
-
       // Fetch the pet module
       let pet = await realm.realmIndexQueryEngine.module(
         new URL(`${testRealm}pet`),
@@ -1568,22 +1728,6 @@ module(basename(__filename), function () {
             @field name = contains(Name);
           }
         `,
-      );
-
-      // Verify initial error state
-      assert.deepEqual(
-        { ...realm.realmIndexUpdater.stats },
-        {
-          instancesIndexed: 0,
-          instanceErrors: 2,
-          moduleErrors: 2,
-          modulesIndexed: 0,
-
-          definitionErrors: 0,
-          definitionsIndexed: 0,
-          totalIndexEntries: 20,
-        },
-        'instance and module are in error state before dependency is available',
       );
 
       // Now create the missing name.gts module
@@ -1641,8 +1785,7 @@ module(basename(__filename), function () {
       );
     });
 
-    // we're going to delete this test soon and it doesn't work with prerender-based definition lookup
-    skip('can incrementally index deleted instance', async function (assert) {
+    test('can incrementally index deleted instance', async function (assert) {
       await realm.delete('mango.json');
 
       let { data: result } = await realm.realmIndexQueryEngine.search({
@@ -1668,12 +1811,11 @@ module(basename(__filename), function () {
       );
     });
 
-    // we're going to delete this test soon and it doesn't work with prerender-based definition lookup
-    skip('can incrementally index instance that depends on updated card source', async function (assert) {
+    test('can incrementally index instance that depends on updated card source', async function (assert) {
       await realm.write(
         'post.gts',
         `
-        import { contains, linksTo, field, CardDef } from "https://cardstack.com/base/card-api";
+        import { contains, linksTo, field, CardDef, Component } from "https://cardstack.com/base/card-api";
         import StringField from "https://cardstack.com/base/string";
         import { Person } from "./person";
 
@@ -1685,6 +1827,12 @@ module(basename(__filename), function () {
               return this.author.firstName + '-poo';
             }
           })
+          static isolated = class Isolated extends Component<typeof this> {
+            <template>
+              <h1><@fields.message/></h1>
+              <h2><@fields.author/></h2>
+            </template>
+          }
         }
       `,
       );
@@ -1696,24 +1844,9 @@ module(basename(__filename), function () {
         },
       });
       assert.strictEqual(result.length, 1, 'found updated document');
-      assert.deepEqual(
-        // we splat because despite having the same shape, the constructors are different
-        { ...realm.realmIndexUpdater.stats },
-        {
-          instancesIndexed: 1,
-          instanceErrors: 1,
-          moduleErrors: 0,
-          modulesIndexed: 1,
-          definitionErrors: 0,
-          definitionsIndexed: 1,
-          totalIndexEntries: 26,
-        },
-        'indexed correct number of files',
-      );
     });
 
-    // we're going to delete this test soon and it doesn't work with prerender-based definition lookup
-    skip('can incrementally index instance that depends on updated card source consumed by other card sources', async function (assert) {
+    test('can incrementally index instance that depends on updated card source consumed by other card sources', async function (assert) {
       await realm.write(
         'person.gts',
         `
@@ -1730,6 +1863,9 @@ module(basename(__filename), function () {
             static embedded = class Embedded extends Component<typeof this> {
               <template><@fields.firstName/> (<@fields.nickName/>)</template>
             }
+            static fitted = class Fitted extends Component<typeof this> {
+              <template><@fields.firstName/> (<@fields.nickName/>)</template>
+            }
           }
         `,
       );
@@ -1741,24 +1877,9 @@ module(basename(__filename), function () {
         },
       });
       assert.strictEqual(result.length, 1, 'found updated document');
-      assert.deepEqual(
-        // we splat because despite having the same shape, the constructors are different
-        { ...realm.realmIndexUpdater.stats },
-        {
-          instancesIndexed: 3,
-          instanceErrors: 1,
-          moduleErrors: 0,
-          modulesIndexed: 3,
-          definitionErrors: 0,
-          definitionsIndexed: 3,
-          totalIndexEntries: 26,
-        },
-        'indexed correct number of files',
-      );
     });
 
-    // we're going to delete this test soon and it doesn't work with prerender-based definition lookup
-    skip('can incrementally index instance that depends on deleted card source', async function (assert) {
+    test('can incrementally index instance that depends on deleted card source', async function (assert) {
       await realm.delete('post.gts');
       {
         let { data: result } = await realm.realmIndexQueryEngine.search({
@@ -1785,9 +1906,9 @@ module(basename(__filename), function () {
             id: `${testRealm}post`,
             isCardError: true,
             additionalErrors: null,
-            message: `${testRealm}post not found`,
+            message: `missing file ${testRealm}post`,
             status: 404,
-            title: 'Not Found',
+            title: 'Link Not Found',
             deps: [`${testRealm}post`],
           },
           'card instance is an error document',
@@ -1795,26 +1916,12 @@ module(basename(__filename), function () {
       } else {
         assert.ok(false, 'search index entry is not an error document');
       }
-      assert.deepEqual(
-        // we splat because despite having the same shape, the constructors are different
-        { ...realm.realmIndexUpdater.stats },
-        {
-          instancesIndexed: 0,
-          instanceErrors: 2,
-          moduleErrors: 0,
-          modulesIndexed: 0,
-          definitionErrors: 0,
-          definitionsIndexed: 0,
-          totalIndexEntries: 23,
-        },
-        'indexed correct number of files',
-      );
 
       // when the definitions is created again, the instance should mend its broken link
       await realm.write(
         'post.gts',
         `
-        import { contains, linksTo, field, CardDef } from "https://cardstack.com/base/card-api";
+        import { contains, linksTo, field, CardDef, Component } from "https://cardstack.com/base/card-api";
         import StringField from "https://cardstack.com/base/string";
         import { Person } from "./person";
 
@@ -1823,9 +1930,15 @@ module(basename(__filename), function () {
           @field message = contains(StringField);
           @field nickName = contains(StringField, {
             computeVia: function() {
-              return this.author.firstName + '-poo';
+              return this.author?.firstName + '-poo';
             }
           })
+          static embedded = class Embedded extends Component<typeof this> {
+            <template><@fields.firstName/> (<@fields.nickName/>)</template>
+          }
+          static fitted = class Fitted extends Component<typeof this> {
+            <template><@fields.firstName/> (<@fields.nickName/>)</template>
+          }
         }
       `,
       );
@@ -1838,299 +1951,6 @@ module(basename(__filename), function () {
         });
         assert.strictEqual(result.length, 1, 'found the post instance');
       }
-      assert.deepEqual(
-        // we splat because despite having the same shape, the constructors are different
-        { ...realm.realmIndexUpdater.stats },
-        {
-          instancesIndexed: 1,
-          instanceErrors: 1,
-          moduleErrors: 0,
-          modulesIndexed: 1,
-          definitionErrors: 0,
-          definitionsIndexed: 1,
-          totalIndexEntries: 26,
-        },
-        'indexed correct number of files',
-      );
-    });
-
-    // Note this particular test should only be a server test as the nature of
-    // the TestAdapter in the host tests will trigger the linked card to be
-    // already loaded when in fact in the real world it is not.
-    test('it can index a card with a contains computed that consumes a linksTo field', async function (assert) {
-      const hassanId = `${testRealm}hassan`;
-      let queryEngine = realm.realmIndexQueryEngine;
-      let hassan = await queryEngine.cardDocument(new URL(hassanId));
-      if (hassan?.type === 'doc') {
-        assert.deepEqual(
-          hassan.doc.data.attributes,
-          {
-            title: 'Untitled Card',
-            nickName: "Ringo's buddy",
-            firstName: 'Hassan',
-            description: null,
-            thumbnailURL: null,
-            cardInfo,
-          },
-          'doc attributes are correct',
-        );
-        assert.deepEqual(
-          hassan.doc.data.relationships,
-          {
-            pet: {
-              links: {
-                self: './ringo',
-              },
-            },
-            'cardInfo.theme': {
-              links: {
-                self: null,
-              },
-            },
-          },
-          'doc relationships are correct',
-        );
-      } else {
-        assert.ok(
-          false,
-          `search entry was an error: ${hassan?.error.errorDetail.message}`,
-        );
-      }
-
-      let hassanEntry = await getInstance(realm, new URL(`${testRealm}hassan`));
-      if (hassanEntry) {
-        assert.deepEqual(
-          hassanEntry.searchDoc,
-          {
-            id: hassanId,
-            pet: {
-              id: `${testRealm}ringo`,
-              title: 'Untitled Card',
-              firstName: 'Ringo',
-              description: null,
-              thumbnailURL: null,
-              cardInfo: {
-                ...cardInfo,
-                theme: null,
-              },
-            },
-            nickName: "Ringo's buddy",
-            _cardType: 'PetPerson',
-            firstName: 'Hassan',
-            title: 'Untitled Card',
-            cardInfo: {
-              theme: null,
-            },
-          },
-          'searchData is correct',
-        );
-      } else {
-        assert.ok(false, `could not find ${hassanId} in the index`);
-      }
-
-      assert.deepEqual(
-        // we splat because despite having the same shape, the constructors are different
-        { ...realm.realmIndexUpdater.stats },
-        {
-          moduleErrors: 0,
-          instanceErrors: 6,
-          modulesIndexed: 10,
-          instancesIndexed: 6,
-          definitionErrors: 0,
-          definitionsIndexed: 10,
-          totalIndexEntries: 26,
-        },
-        'indexed correct number of files',
-      );
-    });
-
-    test('it can index a card with a contains computed that consumes a linksTo field that is NOT in template but uses "isUsed" option', async function (assert) {
-      await realm.write(
-        'task.gts',
-        `
-            import StringField from 'https://cardstack.com/base/string';
-            import {
-              Component,
-              CardDef,
-              contains,
-              field,
-              linksTo,
-            } from 'https://cardstack.com/base/card-api';
-
-
-            export class Team extends CardDef {
-              static displayName = 'Team'
-              @field name = contains(StringField, {isUsed: true});
-            }
-
-            export class Task extends CardDef {
-              static displayName = 'Sprint Task';
-              @field team = linksTo(() => Team, {isUsed: true});
-              @field shortId = contains(StringField, {
-                computeVia: function (this: Task) {
-                  return this.team?.name
-                },
-              });
-
-              //template with no reference to shortId
-              static isolated = class TaskIsolated extends Component<typeof this> {
-              <template>
-              </template>
-              }
-            }
-            `,
-      );
-      await realm.write(
-        'team.json',
-        JSON.stringify({
-          data: {
-            type: 'card',
-            attributes: {
-              name: 'Team B',
-              description: null,
-              thumbnailURL: null,
-            },
-            meta: {
-              adoptsFrom: {
-                module: './task',
-                name: 'Team',
-              },
-            },
-          },
-        }),
-      );
-      await realm.write(
-        'task.json',
-        JSON.stringify({
-          data: {
-            type: 'card',
-            attributes: {
-              title: null,
-              description: null,
-              thumbnailURL: null,
-            },
-            relationships: {
-              team: {
-                links: {
-                  self: './team',
-                },
-              },
-            },
-            meta: {
-              adoptsFrom: {
-                module: './task',
-                name: 'Task',
-              },
-            },
-          },
-        }),
-      );
-      let taskInstance = (await realm.realmIndexQueryEngine.instance(
-        new URL(`${testRealm}task`),
-      )) as any;
-      assert.strictEqual(
-        taskInstance?.type,
-        'instance',
-        'task instance created without any error',
-      );
-    });
-
-    test('sets resource_created_at for modules and instances', async function (assert) {
-      let entry = (await realm.realmIndexQueryEngine.module(
-        new URL(`${testRealm}fancy-person.gts`),
-      )) as { resourceCreatedAt: number };
-
-      assert.ok(entry!.resourceCreatedAt, 'resourceCreatedAt is set');
-
-      entry = (await realm.realmIndexQueryEngine.instance(
-        new URL(`${testRealm}mango`),
-      )) as { resourceCreatedAt: number };
-
-      assert.ok(entry!.resourceCreatedAt, 'resourceCreatedAt is set');
-    });
-
-    test('sets urls containing encoded CSS for deps for a module', async function (assert) {
-      let entry = (await realm.realmIndexQueryEngine.module(
-        new URL('http://test-realm/fancy-person.gts'),
-      )) as { deps: string[] };
-
-      let assertCssDependency = (
-        deps: string[],
-        pattern: RegExp,
-        fileName: string,
-      ) => {
-        assert.true(
-          !!deps.find((dep) => pattern.test(dep)),
-          `css for ${fileName} is in the deps`,
-        );
-      };
-
-      let dependencies = [
-        {
-          pattern: /fancy-person\.gts.*\.glimmer-scoped\.css$/,
-          fileName: 'fancy-person.gts',
-        },
-        {
-          pattern: /test-realm\/person\.gts.*\.glimmer-scoped\.css$/,
-          fileName: 'person.gts',
-        },
-        {
-          pattern:
-            /cardstack.com\/base\/default-templates\/embedded\.gts.*\.glimmer-scoped\.css$/,
-          fileName: 'default-templates/embedded.gts',
-        },
-        {
-          pattern:
-            /cardstack.com\/base\/default-templates\/isolated-and-edit\.gts.*\.glimmer-scoped\.css$/,
-          fileName: 'default-templates/isolated-and-edit.gts',
-        },
-        {
-          pattern:
-            /cardstack.com\/base\/default-templates\/missing-template\.gts.*\.glimmer-scoped\.css$/,
-          fileName: 'default-templates/missing-template.gts',
-        },
-        {
-          pattern:
-            /cardstack.com\/base\/default-templates\/field-edit\.gts.*\.glimmer-scoped\.css$/,
-          fileName: 'default-templates/field-edit.gts',
-        },
-        {
-          pattern:
-            /cardstack.com\/base\/links-to-many-component.gts.*\.glimmer-scoped\.css$/,
-          fileName: 'links-to-many-component.gts',
-        },
-        {
-          pattern:
-            /cardstack.com\/base\/links-to-editor.gts.*\.glimmer-scoped\.css$/,
-          fileName: 'links-to-editor.gts',
-        },
-        {
-          pattern:
-            /cardstack.com\/base\/contains-many-component.gts.*\.glimmer-scoped\.css$/,
-          fileName: 'contains-many-component.gts',
-        },
-        {
-          pattern:
-            /cardstack.com\/base\/field-component.gts.*\.glimmer-scoped\.css$/,
-          fileName: 'field-component.gts',
-        },
-      ];
-
-      dependencies.forEach(({ pattern, fileName }) => {
-        assertCssDependency(entry.deps, pattern, fileName);
-      });
-    });
-
-    test('will not invalidate non-json/non-executable files', async function (assert) {
-      let deletedEntries = (await testDbAdapter.execute(
-        `SELECT url FROM boxel_index WHERE is_deleted = TRUE`,
-      )) as { url: string }[];
-
-      let deletedEntryUrls = deletedEntries.map((row) => row.url);
-
-      ['random-file.txt', 'random-image.png', '.DS_Store'].forEach((file) => {
-        assert.notOk(deletedEntryUrls.includes(file));
-      });
     });
 
     test('should be able to handle dependencies between modules', async function (assert) {
@@ -2324,6 +2144,7 @@ module(basename(__filename), function () {
         'BlogApp module is in resolved module successfully',
       );
     });
+
     test('can write several modules at once', async function (assert) {
       let mapOfWrites = new Map();
       mapOfWrites.set(
@@ -2463,7 +2284,6 @@ module(basename(__filename), function () {
 
       await adapter.remove('test-file.json'); // incremental doesn't get triggered (like in development) here bcos there is no filewatcher enabled
       realm.__testOnlyClearCaches();
-
       let fileExists = await adapter.exists('test-file.json');
       assert.false(fileExists);
       await realm.realmIndexUpdater.fullIndex();
@@ -2496,7 +2316,7 @@ module(basename(__filename), function () {
   });
 
   module('permissioned realm', function (hooks) {
-    setupBaseRealmServer(hooks, matrixURL, { disablePrerenderer: true });
+    setupBaseRealmServer(hooks, matrixURL);
 
     let testRealm1URL = 'http://127.0.0.1:4447/';
     let testRealm2URL = 'http://127.0.0.1:4448/';
@@ -2510,7 +2330,6 @@ module(basename(__filename), function () {
       },
     ) {
       setupPermissionedRealms(hooks, {
-        disablePrerenderer: true,
         // provider
         realms: [
           {
@@ -2560,7 +2379,7 @@ module(basename(__filename), function () {
     module('readable realm', function (hooks) {
       setupRealms(hooks, {
         provider: {
-          [testRealmServerMatrixUserId]: ['read'],
+          ['@node-test_realm:localhost']: ['read'],
         },
         consumer: {
           '*': ['read', 'write'],
